core-controller/zephyr-demo/mock_devices/requirements.txt
python-can>=4.3
rich>=13.7

